[TASK] Disable unapplied TS-highlighting for unknown-variables in CodeMirror
This removes any TS-highlighting for unknown-variables in CodeMirror.
This does not visually change the styling of unknown-variables.

Before a style `other` was defined to be applied to unknown-variables,
but the style `other` does not exist.
Thereby no style was applied and the unknown-variables were already
styled in default-text-style 'black'.

Before a warning was shown in browser's dev console:
`Unknown highlighting tag other`

Resolves: #105521
Releases: main, 13.4, 12.4
Change-Id: I389791b2d6c7f51caafc552278468acea4321654
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86999
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>