[BUGFIX] Fix shorthand translation key handling in extbase validators
Since #102326 it is possible to customize extbase validation error
messages on validator basis.

With the change, all extbase validators were changed to use the
`LLL:` syntax for validation messages, adding an early return to
`translateErrorMessage()` in `AbstractValidator`, in order to
prevent messages not beginning with `LLL:` from being translated.

However, this did not respect a common scenario where the
`translateErrorMessage()` method was called with a shorthand
translation key (e.g. `validator.foo.message`) and an extension
name as second argument.

In this case, extbase `LocalizationUtility` would always use the
`locallang.xlf` file of the given extension name to lookup
possible translation keys.

This change fixes the problem by ensuring that
`translateErrorMessage()` always calls
`LocalizationUtility::translate()`, if an extension name is provided
as second argument to the method.

Tests are added to validate this behaviour.

Resolves: #105536
Related: #102326
Releases: main, 13.4
Change-Id: If47e7e1aa0f178fd3aeb98c6b5fd383ef3a8c05e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86877
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>