[BUGFIX] Add FormEngine transformation for dbType=datetime in flexforms
The database-to-backend-from transformation for dbType=datetime fields
(which are supposed to be stored in native databse DATETIME fields,
but are stored as datetime strings in flexforms) were not configured.
As a consequence the raw value (not an ISO8601 string, but Date + Time
separated by a space-char) was passed into formengine and incorrectly
interpreted as timestamp, because it failed to be parsed as a ISO8601
date.

Add this configuration and remove the previously erroneously used
timestamp parsing which is no longer needed since #104309 and was
only implicitly used to transform the datetime, which hid
this bug by resolving into an incorrectly created date.
For this invalid case we'll now throw an error instead of hiding
this case with an invalid transformation.

Note that this cleanup is related to #104309, but the issue has
been present before. Prior to #104309 the field was rendered
as "Invalid DateTime", while the switch to ISO8601 in #104309
resulted in an invalid date to be generated:
A flexform value like "1960-01-01 05:42:23" was interpreted as 1960
seconds (unix timestamp) after 1970 which results in a fake-date to be
calculated: 1970-01-01T00:32:40Z (32:40 because 1960s = 32 * 60 + 40).

Releases: main, 13.4, 12.4
Resolves: #105550
Related: #104309
Change-Id: Ibbf4b4e709358247d6eb8d006bedb05818f1d9c2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87068
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>