[BUGFIX] Fix rendering of native datetime fields as record title
FormDataProvider\DatabaseRowDateTimeFields misinterprets the native
LOCALTIME datetime value as UTC and creates an invalid ISO8601
datetime string in the current server timezone based on the incorrectly
assumed UTC base value, which means the time value has an effective
timezone offset that matches the current server timezone.

As a backportable workaround the timezone offset is now normalized back
to LOCALTIME, as a change to DatabaseRowDateTimeFields would be
considered breaking.
This ugly workaround will be removed with a refactoring of
DatabaseRowDateTimeFields to use unqualified ISO8601 localtime
in #105549.

Resolves: #103693
Related: #105549
Releases: main, 13.4, 12.4
Change-Id: I621658f1aadca3ae6425560909c1e27bebb2ec25
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86917
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>