[BUGFIX] Display correct method for callable event listeners
In case an event listener is not registered
with a `method`, it's treated as callable
by the ListenerProvider and therefore its
`__invoke()` method is called.

For such cases, the configuration module
displayed `method =`, which is misleading
as this might look like the listener
registration is broken.

Therefore, the configuration module now
displays `method = __invoke` in case a
listener is used as callable (does not
define a  dedicated method in the
registration).

Resolves: #105568
Releases: main, 13.4, 12.4
Change-Id: Ib343629087d830c9747157886b681c65b663cb70
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86938
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>