[DOCS] Hint LanguageService->sL() docs on placeholder usage
The inline code comment is used for rendering the documentation
at [1]. It currently is missing information that LanguageService->sL()
is not doing any placeholder interpolation, and that this needs
to be performed individually.

People may lean towards using Extbase LocalizationUtility in
a wrong context, because that offers placeholder substitution.

[1] https://docs.typo3.org/permalink/t3coreapi:LanguageService-api

Resolves: #105587
Releases: main, 13.4, 12.4
Change-Id: I7f86a43b9d68bee865d899e6bab9e1bf6f4ec84c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86979
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Lina Wolf <112@linawolf.de>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>