[TASK] Provide expression with TypoScript condition errors
TypoScript conditions are commonly tailored to frontend usage and often
miss guards like "traverse()" to avoid accessing undefined objects or
properties. In this case an unspecific RuntimeException was thrown which
mentioned the error but made it hard to understand the reason/context.

We now throw a more specific RuntimeException to mention the TypoScript
condition context and the faulty expression.

The same is done for the error log message in case of TypoScript
condition syntax errors.

Resolves: #105191
Releases: main, 13.4
Change-Id: If4238297f8c24f7e2cd93403bd440190816b79db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88440
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>