[BUGFIX] Prevent PHP errors for empty Fluid templates
Because of the internal Fluid implementation, Fluid's `render()`
method does not always return a string. If a template only contains
an integer, an integer is returned, if only a ViewHelper is called that
returns an array, an array is returned. Most importantly, empty
template files return `null` instead of an empty string.

For TYPO3's Fluid adapter, this behavior doesn't make sense. Thus,
the result is now cast to string, which prevents PHP errors in the most
common use cases.

Resolves: #105588
Releases: main, 13.4
Change-Id: I8385b562e286b0b1b41650b2fe7e7f67fb4d12bc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87010
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>