[BUGFIX] Handle db value null in AbstractItemProvider
While "item-able" field types do not support null values
in the backend, it is still possible null was added via
external API (e.g. extbase).

A guard clause is added for null and will return an empty
array if encountered. This is now needed, as GU:trimExplode
is strict since TYPO3 v13.

In future, it could be considered to introduce option `nullable`
for these field types as well. This would be especially useful
for optional relations and would erase the need to define a
default item like "Please select...".

Resolves: #105619
Related: #101453
Releases: main, 13.4
Change-Id: Ic35b1c111c4cb617e6850f34046d483af0c1c13c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87180
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>