[BUGFIX] Resolve RecordPropertyClosure for flex fields
This introduces a new model class `FlexFormFieldValues`,
which represents a record's flex field values. By
implementing `ContainerInterface` and `ArrayAccess`,
it allows to access the property values in fluid
templates with various available syntaxes. Most
importantly, in case a flex field value is a
`RecordPropertyClosure`, it's now resolved properly.
This works recursively and the resolved value is
"persisted" so does not need to be resolved again.

Support for fields with dot syntax, e.g. "some.field"
is implemented as well.

Due to `ContainerInterface` and `ArrayAccess`,
accessing such property is possible via:

- $field['some']['field']
- $field->get('some')['field']
- $field->get('some.field')

Resolves: #105621
Releases: main, 13.4
Change-Id: I0ce443a5a02fc5a19cc80163c511e06ce68b8f52
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87083
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>