[BUGFIX] Mark file-storage caching as active on selection in install tool
In "Admin Tools -> Settings -> Configuration Presets: Cache Settings"
a previously selected entry `Prefer File storage for caching` is marked
as active now.

Before a selection of `File storage` resulted in `Custom configuration`
being marked as active on relaod of the `Configuration Presets`-modal.

This also fixes the test-case: The test checks if the radio-button
which is selected has the value `File`.
Before the test only checked if a radio-button having value `File`
exists, but didn't check if this radio-button was actually selected.

Resolves: #105637
Releases: main, 13.4, 12.4
Change-Id: I197599689cfe8c4ca19e296e06df0b400b4f6185
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87566
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>