[TASK] FormEngine: do not return a value in `validateField()`
The FormEngine validation method `validateField()` previously returned a
value, which does not make much sense in the method's scope.

There is an actual value modification going on, but the code copied the
value into a separate variable, which was returned. In simple words:
input === output.

This is even more questionable and revealed a no-op branch regarding
FlexForm sections, which has been removed as well.

Resolves: #105656
Releases: main
Change-Id: I426c30cac49bdc337b6a6733ac7d25dbe3a1353b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86385
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Kienast <akienast@scripting-base.de>