[TASK] Mitigate postgres DB issue in styleguide
Command
> Build/Scripts/runTests.sh -s acceptance -d postgres -i 15 -p 8.3
currently fails since #105595.
It seems either DefaultTcaSchema is wrong with the generated
int field, or DataHandler should avoid writing int 0 when
no row is selected. This needs further analysis. For now,
we create an ext_tables.sql for select_tree_7, that makes
the field a string to prevent red nightlies and gain
some time for further debug of this rather seldom TCA
configuration in combination with postgres.

Resolves: #105660
Related: #105595
Releases: main
Change-Id: Ic73daf8541458af785a39df159bf55f81d46b7dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87109
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>