[BUGFIX] Preserve correct line breaks in `<textarea>`
The HTML `<textarea>` element has very specific rules for leading
linefeed (0x0a) characters: if the first char of the content is a
linefeed, it is to be ignored by parsers. [1]

To represent the exact number of leading line breaks, a supplementary
linefeed character needs to be prepended to the textarea value, which
will always be ignored, but ensures that subsequent linefeeds are
respected.

[1] https://html.spec.whatwg.org/multipage/parsing.html#parsing-main-inbody:~:text=A%20start%20tag%20whose%20tag%20name%20is%20%22textarea%22

Resolves: #94516
Releases: main, 13.4
Change-Id: Icc3f976e01f1e12932e8d9f24898daf7901c99cb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87650
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>