[TASK] Extend lifetime of `tableoptions` and `collate` deprecation
Since TYPO3 v13 the `tableoptions` array and `collate` sub-key has
been deprecated and triggers deprecation notices when migrated to
the new options.

Originally scheduled to remove the migration with TYPO3 v14 following
casual deprecation policiy, it has been decided to extend the lifetime
of the deprecation and migration to TYPO3 v15 or later as it does not
hurt to have the migration and deprecation in place.

This change modifies the code comments and depreceation changelog
entry to reflect the lifetime extending decision.

Resolves: #105495
Related: #105377
Related: #105506
Related: #105297
Releases: main, 13.4
Change-Id: Icd33586f186d133ac256773f47f9a504fc9f6f23
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86812
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>