[TASK] Remove superfluous radio fields in extension manager
The `DownloadExtensionViewHelper` in extension manager outputs
two radio fields with options to install TER extensions either
as local or system extension. The radio fields are however
hidden via css and the functionality to install TER extensions
as a system extension has been removed with #102897.

This change removes the output of the radio fields in the
`DownloadExtensionViewHelper` and usage of the fields in
JavaScript.

Resolves: #105715
Related: #102897
Releases: main, 13.4
Change-Id: I78e968e6c5d6a999ef72e292c22ff94471865d05
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87268
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>