[TASK] Use default `redirects.hitCount` setting in acceptance test
Composer mode acceptance tests has been introduced with #103297,
including a conditional check for the `redirects.hitCount` based
on the environment (classic or composer mode) required due to the
fact that typo3/testing-framework sets this toggle to `true`.

Valid request has been made to remove the toggle setting from the
typo3/testing-framework setup [1].

This change removes the codeception conditional based checkbox check
and resets the `redirects.hitCount` toggle to the default value as a
temporary solution.

After removing enabling this toggle in the testing-framework and
relasing this, the temporary solution can be removed again.

[1] https://github.com/TYPO3/testing-framework/issues/554

Resolves: #105725
Related: #103297
Releases: main, 13.4
Change-Id: I610951d211ead78cd17567a4dde9f9d6d98ae7d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87280
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>