[BUGFIX] Include modal-trigger button value in submitted form data
An intercepted (via confirm modal) form submit button with a name and
value attribute is now included in the form-payload by requesting the
submit with the button as context instead of manually submitting the
form without trigger-button context.

This ensures that button values are actually submitted and can be used
to carry relevant payload. The site configuration overview is
adapted to make use of this by removing unnecessarily duplicated forms.

Codewise we now make use of the read-only `form` property of HTML
elements that are associated to a form (either implicitly via nesting
or explicit via `form=""` attribute) to avoid manual DOM lookups, see:
https://developer.mozilla.org/en-US/docs/Web/API/HTMLObjectElement/form

Releases: main, 13.4, 12.4
Resolves: #105758
Related: #96542
Change-Id: I3f4486b3c89ea9447f751b5f6a1fcaa0020f3d0c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87404
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>