[TASK] Make FormEngine InlineStackProcessor stateless
FormEngine related InlineStackProcessor is quite a
convoluted beast that never received true love, not
even during FormEngine rewrite in v7/v8.

As one key issue, the service is stateful since it
parks the given "inlineStructure" in a class property
that has to be "initialized" and is then used in all
read and manipulation methods.

Shared instances of this class are already injected at
some places within FormEngine, making its statefulness
even more unfortunate.

The patch puts an end to this by changing the methods to
always receive the current "inlineStructure" as argument,
removing the state property and declaring the class
readonly. Consumers are carefully adapted and instances
of this now stateless service are now always injected.

Resolves: #105798
Releases: main
Change-Id: Ic29c2dc4c68f0cb310fa107e3184fe7410275f31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87405
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>