[BUGFIX] Ensure Indexed Search can handle PPTX and XLSX files
Thanks to Xavier Perseguers for providing this fix
which properly indexes XLSX and PPTX files.

Problem for XLSX was that the wrong "unzipped" file was indexed,
a file that contains some kind of metadata which are basically only
pointers (integers). The content extracted is plain wrong and useless.

Problem for PPTX was that only the content of Slide 1 was extracted,
it missed content from all other slides.

Resolves: #99527
Releases: main, 13.4
Change-Id: I8293ad8a247b243bec80188471ee2cee3a5151f7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87441
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>