[BUGFIX] Fix wrong sorting and l10n_source for copied pages
First off: Huge thanks to Denis Mir who tracked down
the actual root of the problem.

Right now, there are various scenarios where
"l10n_source" and "l10n_parent" of pages are not the
same in a running installation, which is technically
not possible currently.

However, this happens when a page is copied with their
subpages, and the actual pages to be copied are selected.
In this case, TYPO3 had a bug which was caused by
incorrectly querying the actual pages to copy.

Since "pages" always need a record in the default language
(sys_language_uid=0), they are the ones to be copied, and
the translations are copied along.

However, the query was based to fetch also all translations
of a page as well, which share the same PID and sorting,
thus, resulting in different orderings (sorting) AND
thus, the l10n_source was wrongly filled.

This change finally fixes this issue properly.

In order for you to make sure this works on your machine,
EXT:dbdoctor should help. Or, if you aren't on a large
and sophisticated project without workspaces, this one
could also do the trick:

  "UPDATE pages SET l10n_source=l10n_parent WHERE sys_language_uid>0;"

Also please note that some sorting in translated records might
be wrong.

Resolves: #105765
Resolves: #91752
Resolves: #97469
Releases: main, 13.4, 12.4
Change-Id: Ib9af3bc84060f85a254f9347d4f8b23d8f46f81c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87452
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>