[BUGFIX] Remove redundant overrideChildTca from image fields
Since #101277 the behavior for file fields was changed to have the full
palette (imageOverlayPalette, audioOverlayPalette, ...) per default.
Now, the old overrideChildTca definitions for image fields can be
removed, greatly simplifying the TCA definition.

Some fields defined the imageOverlayPalette for type "unknown" and
"text" (see #79622, #56884 and #33752). This is obviously a
misconfiguration and will lead to an unusable crop field. The fields
"alternative" and "link" don't make sense either for type text.

For backwards compatibility, this patch adds an intermediate palette
called "extendedBasicOverlayPalette", which keeps the "alternative" and
"link" fields for type unknown, text and application. This workaround
should be removed again in v14.

Styleguide can be safely changed, as it's not API.

Resolves: #105823
Related: #101277
Related: #79622
Related: #56884
Related: #33752
Releases: main, 13.4
Change-Id: Idd83cf5101aa36df9a06c2f214d10efee13e91f7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87484
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>