[BUGFIX] Respect aspect ratio for thumbnails in filelist and clipboard
When uploading image files to the file list, the thumbnails are
initially displayed with their correct aspect ratios. If the user
reloads the file list/folder list view, the thumbnails were displayed
in a square shape, which sometimes led to a distorted display.

The patch now uses a maximum width/height instead of a fixed
width/height to display the thumbnails with their correct aspect
ratios, analogous to their initial display after uploading.

Since the clipboard display was also affected, the maximum width/height
setting was also applied to it.

Resolves: #105463
Releases: main, 13.4
Change-Id: I0211f27e0f9947408c3cb6d5714aebe71e9ce89a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87485
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <gh@faktor-e.de>