[BUGFIX] Add border on whole cropper face
The cropper face, indicating the current image sub-part used when
processing the image, now has a border around the whole canvas and not
only around the orientation markers, whose 50% transparency has been
removed as well.

The `opacity` of the cropper face is replaced with a background color
utilizing the alpha channel to make the border around the whole cropper
face visible.

Resolves: #105831
Releases: main, 13.4
Change-Id: Ifeecc9124a91a5fddf81c11e56930ebf441d680b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87523
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>