[BUGFIX] Update CKEditor model in source edit view on input
This patch fixes a bug where changes made in "Source" editing mode
would not be checked during FormEngine validation, mistakenly allowing
to save records when an RTE field is marked as 'required'.

Our CKEditor integration allows to edit the editor document's source
code by default by toggling the "Source" button in the UI. Originally,
the editor model is updated once either the document is saved or the
user disables source editing again.

This has the disadvantage that FormEngine cannot intercept any changes
in an early state, e.g. marking a field as "changed". If source editing
gets enabled, an event listener for the `input` event is now registered
on the textarea that is rendered by source editing, which we have to
pick by ourselves as it is considered private API. The custom event
handler updates the editor model on every input. Since this might cause
performance bottlenecks, the update is debounced every few milliseconds.

Since we deal with private API which might break in future CKEditor
updates, an error is explicitly thrown if the textarea cannot be found.

Resolves: #105701
Releases: main, 13.4, 12.4
Change-Id: Idef8db47772a85205a79e263b517f715393d1213
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88480
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>