[FEATURE] Add search by page URI in page tree filter and livesearch
The backend page tree can now be filtered by entering one or more
frontend URIs, similiar to the search by page UID.

It can be combined with the other search methods by page/nav title
or page uid.

It uses the existing page/workspace/permission checks, and resolves
the URI using the existing SiteMatcher/PageRouter API (via a new
intermediate internal helper class `SiteUrlResolver`).

The filter is implemented through the existing PSR-14 event
`BeforePageTreeIsFilteredEvent`.

Additionally, the same functionality is provided to the Live Search,
utilizing a new PSR-14 event `ModifyConstraintsForLiveSearchEvent`,
which was added because `ModifyQueryForLiveSearchEvent` did not allow
to influence the OR-constraints. The event is dispatched from both
the `PageRecordProvider` and the `DatabaseRecordProvider`, and the
event listener `AddLiveSearchFrontendUriResolverListener` acts on this
on the `PageRecordProvider` only. Database record's do not have
frontend URIs. The dispatched event however can be used for other listener
implementations in the future.

Similar to the translation page search, both a User TSConfig
setting and a persisted user setting control whether this feature
is enabled (default setting):

  options.pageTree.searchByFrontendUri = 1

A runtime cache is used to store matched frontend URI
uids so that labels with hints at the match reasons can be
embedded.

Matches in translated pages are specifically labelled in the
page tree result, and live search yields two entries for
both the translated page as well as the default language
variant of it.

(Please note there is a seperate bug that currently multilingual
results in the live search do not point/link to the proper
target backend functionality. This is unrelated and needs
to be fixed seperately)

Resolves: #105827
Related: #105833
Related: #107961
Releases: main
Change-Id: I0dbdc9334222216542393fa586932b2ff2ce57eb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87549
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Sven Liebert <mail@sven-liebert.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>