[BUGFIX] Make install command accept create-site 'no'
When you run the setup command without parameter you are informed
that the default for "create a basic site" is "no." However using
"no" as value for either the argument `--create-site` or the
enviroment variable `TYPO3_SETUP_CREATE_SITE=no` currently causes
errors:

In SetupCommand.php line 626:

The given url for the site name is not valid! Please try again.

Instead of "normalizing" just the user input of the command prompt,
all possible input ways (prompt, argument, variable) the same way by
leaving the interpretation of the input to the validator.

Resolves: #105871
Releases: main, 13.4
Change-Id: Ic836984eecea7ad45356d24eab7bdaabfa4476ea
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87624
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>