[BUGFIX] Align date picker format with BackendUtility::datetime()
The method `BackendUtility::datetime()` renders dates in a format where
the time comes after the date, basically "ddmmyy hhmm". The format used
in the date picker is swapped for historic reasons, causing a slight
inconsistency in the backend UI.

The format is now aligned with the rest of the backend to ensure having
a consistent UI. Also, it feels more natural to enter the date first,
then the time.

Resolves: #85002
Releases: main, 13.4
Change-Id: I35e95139d4e8e3ad484cd3a5d5dba54663eebae1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87665
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>