[TASK] Ensure decorated `Connection->transactional()` method works
TYPO3 decorates the `\Doctrine\DBAL\Connection` class with the custom
implementation `\TYPO3\CMS\Core\Database\Connection` and uses is as
the default connection managed by the `ConnectionPool`.

The `transactional()` method is not used by TYPO3 directly yet while
extension developers have a valid interest in a working state of the
simplified transaction handling.

This change adds some tests to ensure that simplified `transactional()`
handling works and adds the method to the decoracted class along with
a inline phpdocblock override to satisfy PHPStan revealed by the tests.

Resolves: #104561
Releases: main, 13.4
Change-Id: If55b7d6e4f706971e810d44de0431f9343e305c2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87693
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>