[TASK] Switch to `selenium/standalone-chromium`
Replace the CI image `selenium/standalone-chrome` with
`selenium/standalone-chromium`, which works on multiple platforms.
This makes the architecture check for arm64 obsolete.

Due to a window size related bug in the new headless mode in
chromium v128, we're pinning to chromium v131 (chromium v132 drops the
old headless mode). We'll try to upgrade to newer version once the
related chromium bug is resolved:
https://issues.chromium.org/issues/362522328

Releases: main, 13.4, 12.4
Resolves: #106312
Change-Id: Ie48d3bd1f708f26db9b31abfbe12da7e505448a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88503
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>