[TASK] Refine styling of color-scheme toggle button
The button is now always displayed with a default button
styling and no longer uses a primary styling in dark mode.

The original intention was to toggle light and dark mode.
An active state would have been fine if we did not change
the label and it simply stated "enable/disable." However,
we are showing the currently selected value here. For this
reason, the primary state is invalid and behaves inconsistently
with other elements in the backend.

To better indicate that we are switching a state here,
another icon has been added to the button. This icon
helps to represent the triggered action alongside the
currently active mode.

Resolves: #105924
Releases: main, 13.4
Change-Id: I6849d3333fa014cedefdcbbb999b21579a58c2db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87733
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>