[SECURITY] Circumvent parser deviation in PSR-7 URI object
Adapt the URI object to refuse construction based on invalid URL
inputs.

As `parse_url()` is defined to not validate the resulting parts of
the parsed URL, a parser deviation between PHP and browser URL
implementations is possible when URL parts contain invalid characters.

The missing validation is now performed via PHP upstream rules built
into `filter_var()`. Note that `filter_var` has two limitations:
It requires a fully qualified URL and it can not validate IDN domain
names that have not been converted to ascii before.
In order to circumvent these limitations we'll imply a intermediate fake
default-scheme and perform IDN-to-ascii conversion for the purpose
of the validation.

Resolves: #105170
Releases: main, 13.4, 12.4
Change-Id: I63d02de49de53513a8f3b4442a35dccb01b92eff
Security-Bulletin: TYPO3-CORE-SA-2025-002
Security-References: CVE-2024-55892
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87737
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>