[BUGFIX] Add missing minitems/maxitems EXT:styleguide generators
The EXT:styleguide generators for 'selectTree' and 'file' TCA
types are now able to evaluate 'minitems' and 'maxitems' attributes
for tree creation.

With this, the created 'required' example data data will no longer be
saved in a state that has initial validation failures when editing
the dummy records.

This is a pre-patch for #105940.

Resolves: #105947
Related: #105940
Releases: main, 13.4
Change-Id: I9795258e6cb378df973dba7dade4f6e05fc2b658
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87778
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>