[BUGFIX] Remove ColumnDiff overwrite to support doctrine/dbal 4.2.2
The doctrine `Doctrine/DBAL 4.2.2` [1] release includes a change to
consider column platform options as changed to support PostgreSQL
`JSONB`.

The newly introduced method `ColumnDiff::hasPlatformOptionsChanged()`
tries to access the private properties `newColumn` and `oldColumn`
which are not accessible in the TYPO3 extended `ColumnDiff` class.
This extended class replaced these private properties by publicly
modifiable ones. This lead to a TypeError during the database compare
for PostgreSQL with Doctrine DBAL 4.2.2.

To mitigate this issue object composition instead of object
modification is used to apply quoting to the ColumnDiff oldColumn
without the need for a custom ColumnDiff implementation.

Resolves: #105959
Releases: main, 13.4
Change-Id: Ie07d373cd92ce4cb535341e140d3440daeac7174
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87801
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>