[BUGFIX] Re-allow saving nullable \DateTime properties as null
If an entity property is a nullable \DateTime, which is also marked
nullable via TCA and which is stored as a DATETIME in the database,
a nulled property must be stored as null.

Currently, all null values are converted to 0000-00-00 00:00:00, which
is not only invalid but also not the desired value.

This change restores the former behaviour.

Releases: main, 13.4
Resolves: #105966
Related: #103641
Change-Id: Iebbafdce59acaa0c5595d30f54350ae2c7e732ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87814
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>